+++ RIAssigner/cli/LoadDataAction.py
@@ -1,6 +1,5 @@
 import argparse
 
-from RIAssigner.utils import get_extension
 from RIAssigner.data import MatchMSData, PandasData
 
 
@@ -9,10 +8,12 @@ class LoadDataAction(argparse.Action):
         super().__init__(option_strings, dest, **kwargs)
 
     def __call__(self, parser, namespace, values, option_string=None):
-        filetype = get_extension(values)
-        if filetype == '.msp':
-            data = MatchMSData(values)
-        elif filetype in ['.csv', '.tsv']:
-            data = PandasData(values)
+        filename = values[0]
+        filetype = values[1]
+        rt_unit = values[2]
+        if filetype == 'msp':
+            data = MatchMSData(filename, filetype, rt_unit)
+        elif filetype in ['csv', 'tsv']:
+            data = PandasData(filename, filetype, rt_unit)
 
         setattr(namespace, self.dest, data)