+++ tests/builders/MatchMSDataBuilder.py
@@ -6,6 +6,7 @@ class MatchMSDataBuilder:
     def __init__(self):
         self.filename = None
         self._rt_unit = 'seconds'
+        self._filetype = "msp"
 
     def with_filename(self, filename: str):
         self.filename = filename
@@ -15,5 +16,9 @@ class MatchMSDataBuilder:
         self._rt_unit = rt_unit
         return self
 
+    def with_filetype(self, filetype: str):
+        self._filetype = filetype
+        return self
+
     def build(self) -> MatchMSData:
-        return MatchMSData(self.filename, self._rt_unit)
+        return MatchMSData(self.filename, self._filetype, self._rt_unit)