+++ tests/test_data_MatchMSData.py
@@ -5,8 +5,8 @@ import pytest
 from matchms.exporting import save_as_msp
 from matchms.importing import load_from_msp
 
-from tests.builders.MatchMSDataBuilder import MatchMSDataBuilder
-
+from tests.builders import MatchMSDataBuilder
+from RIAssigner.utils import get_extension
 here = os.path.abspath(os.path.dirname(__file__))
 testdata_dir = os.path.join(here, 'data', 'msp')
 
@@ -86,7 +86,8 @@ def test_equal(filename):
     expected = MatchMSDataBuilder().with_filename(filename).build()
 
     assert expected == actual
-    
+
+
 @pytest.mark.parametrize("filename", ["PFAS_added_rt.msp", "recetox_gc-ei_ms_20201028.msp"])
 def test_basic_write(filename, tmp_path):
     # TODO: Reafactor with load_test_file
@@ -108,3 +109,5 @@ def test_basic_write(filename, tmp_path):
         actual = file.readlines()
 
     assert expected == actual
+
+